requirements.txt
aiofiles==25.1.0
anyio==4.12.1
argcomplete==3.6.3
async-lru==2.1.0
attrs==25.4.0
backports.zstd==1.3.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
boto3==1.42.30
botocore==1.42.30
brotli==1.2.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cross-web==0.4.1
cryptography==46.0.3
cycler==0.12.1
dacite==1.9.2
Deprecated==1.3.1
dill==0.4.1
dnspython==2.8.0
docutils==0.22.4
fiftyone==1.11.1
fiftyone-brain==0.21.4
fiftyone_db==1.4.1
filelock==3.20.3
filetype==1.2.0
fonttools==4.61.1
fsspec==2026.1.0
ftfy==6.3.1
glob2==0.7
graphql-core==3.2.7
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
Hypercorn==0.18.0
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
inflate64==1.0.4
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
Kivy==2.3.1
Kivy-Garden==0.1.5
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.8
mongoengine==0.29.1
motor==3.6.1
mpmath==1.3.0
multivolumefile==0.2.3
narwhals==2.15.0
ndiff==7.96
networkx==3.6.1
numpy==2.4.1
opencv-python==4.13.0.90
opencv-python-headless==4.13.0.90
packaging==25.0
pandas==2.3.3
paramiko==3.5.1
pillow==10.4.0
plotly==6.5.2
polars==1.37.1
polars-runtime-32==1.37.1
pprintpp==0.4.0
priority==2.0.0
psutil==7.2.1
py7zr==1.1.2
pybcj==1.0.7
pycparser==2.23
pycryptodomex==3.23.0
pydash==8.0.6
Pygments==2.19.2
pymongo==4.9.2
PyNaCl==1.6.2
pyparsing==3.3.1
pyppmd==1.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
rarfile==4.2
regex==2026.1.15
requests==2.32.5
retrying==1.4.2
rtree==1.4.1
s3transfer==0.16.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.9.0
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.8.2
sse-starlette==0.10.3
sseclient-py==1.9.0
starlette==0.52.1
strawberry-graphql==0.289.0
sympy==1.14.0
tabulate==0.9.0
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2026.1.14
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ultralytics==8.4.6
ultralytics-thop==2.0.18
universal-analytics-python3==1.1.1
urllib3==2.6.3
voxel51-eta==0.15.2
wcwidth==0.2.14
wheel @ file:///opt/homebrew/Cellar/python%403.13/3.13.2/libexec/wheel-0.45.1-py3-none-any.whl#sha256=b9235939e2096903717cb6bfc132267f8a7e46deb2ec3ef9c5e234ea301795d0
wrapt==2.0.1
wsproto==1.3.2
xmltodict==1.0.2
